+++ src/main/scala/com/dataintuitive/viash/ViashConfig.scala
@@ -18,9 +18,14 @@
 package com.dataintuitive.viash
 
 import com.dataintuitive.viash.config.Config
-import java.nio.file.{Paths}
+import com.dataintuitive.viash.helpers.IO
+
+import java.nio.file.{Files, Paths}
 import io.circe.syntax.EncoderOps
 import io.circe.yaml.Printer
+
+import scala.sys.process.Process
+import com.dataintuitive.viash.platforms.DebugPlatform
 
 object ViashConfig {
   private val printer = Printer(
@@ -51,22 +56,17 @@ object ViashConfig {
       println("Could not find a main script in the Viash config.")
       System.exit(1)
     }
-    val mainScript = fun.mainScript.get
-
     // check if we can read code
-    val maybeCode = mainScript.readWithPlaceholder(fun)
-    if (maybeCode.isEmpty) {
+    if (fun.mainScript.get.read.isEmpty) {
       println("Could not read main script in the Viash config.")
       System.exit(1)
     }
-    val code = maybeCode.get
-
     // check if main script has a path
-    if (mainScript.uri.isEmpty) {
+    if (fun.mainScript.get.uri.isEmpty) {
       println("Main script should have a path.")
       System.exit(1)
     }
-    val uri = mainScript.uri.get
+    val uri = fun.mainScript.get.uri.get
 
     // check if main script is a local file
     if (uri.getScheme != "file") {
@@ -75,14 +75,25 @@ object ViashConfig {
     }
     val path = Paths.get(uri.getPath())
 
-    // create new script
-    val newScript = mainScript.copyResource(
-      text = Some(code),
-      dest = Some(mainScript.filename),
-      path = None
-    )
+    // debugFun
+    val debugPlatform = DebugPlatform(path = uri.getPath())
+    val debugFun = debugPlatform.modifyFunctionality(fun)
 
-    // write to destination
-    newScript.write(path, overwrite = true)
+    // create temporary directory
+    val dir = IO.makeTemp("viash_inject_" + config.functionality.name)
+    println(s"Running injection in temporary directory: '$dir'")
+
+    // build regular executable
+    Files.createDirectories(dir)
+    IO.writeResources(debugFun.resources.getOrElse(Nil), dir)
+
+    // run command, collect output
+    val executable = Paths.get(dir.toString, fun.name).toString
+    val exitValue = Process(Seq(executable), cwd = dir.toFile).!
+
+    println("LALALA " + debugPlatform.path)
+    println("Exit value: " + exitValue)
   }
+
+
 }