+++ src/main/scala/com/dataintuitive/viash/platforms/package.scala
@@ -33,8 +33,9 @@ package object platforms {
   implicit val encodeNativePlatform: Encoder.AsObject[NativePlatform] = deriveConfiguredEncoder
   implicit val decodeNativePlatform: Decoder[NativePlatform] = deriveConfiguredDecoder
 
-  implicit val encodeDebugPlatform: Encoder.AsObject[DebugPlatform] = deriveConfiguredEncoder
-  implicit val decodeDebugPlatform: Decoder[DebugPlatform] = deriveConfiguredDecoder
+  // there is no reason to parse debug platforms from yaml
+  // implicit val encodeDebugPlatform: Encoder.AsObject[DebugPlatform] = deriveConfiguredEncoder
+  // implicit val decodeDebugPlatform: Decoder[DebugPlatform] = deriveConfiguredDecoder
 
   implicit def encodePlatform[A <: Platform]: Encoder[A] = Encoder.instance {
     platform =>
@@ -43,7 +44,7 @@ package object platforms {
         case s: DockerPlatform => encodeDockerPlatform(s)
         case s: NextFlowPlatform => encodeNextFlowPlatform(s)
         case s: NativePlatform => encodeNativePlatform(s)
-        case s: DebugPlatform => encodeDebugPlatform(s)
+        // case s: DebugPlatform => encodeDebugPlatform(s)
       }
       objJson deepMerge typeJson
   }
@@ -54,7 +55,7 @@ package object platforms {
         cursor.downField("type").as[String] match {
           case Right("docker") => decodeDockerPlatform.widen
           case Right("native") => decodeNativePlatform.widen
-          case Right("debug") => decodeDebugPlatform.widen
+          // case Right("debug") => decodeDebugPlatform.widen
           case Right("nextflow") => decodeNextFlowPlatform.widen
           case Right(typ) => throw new RuntimeException("Type " + typ + " is not recognised.")
           case Left(exception) => throw exception