+++ src/main/scala/com/dataintuitive/viash/wrapper/BashWrapper.scala
@@ -85,7 +85,8 @@ object BashWrapper {
   def wrapScript(
     executor: String,
     functionality: Functionality,
-    mods: BashWrapperMods = BashWrapperMods()
+    mods: BashWrapperMods = BashWrapperMods(),
+    debugPath: Option[String] = None
   ): String = {
     val mainResource = functionality.mainScript
 
@@ -109,23 +110,48 @@ object BashWrapper {
 
     // DETERMINE HOW TO RUN THE CODE
     val executionCode = mainResource match {
+      // if mainResource is empty (shouldn't be the case)
       case None => ""
+
+      // if mainResource is simply an executable
       case Some(e: Executable) => " " + e.path.get + " $VIASH_EXECUTABLE_ARGS"
-      case Some(res) =>
+
+      // if mainResource is a script
+      case Some(res) if debugPath.isEmpty =>
         val code = res.readWithPlaceholder(functionality).get
         val escapedCode = escapeViash(code)
+
+        // check whether the script can be written to a temprorary location or
+        // whether it needs to be a specific path
+        val scriptSetup = 
+          s"""
+            |tempscript=\\$$(mktemp "$$VIASH_TEMP/viash-run-${functionality.name}-XXXXXX")
+            |function clean_up {
+            |  rm "\\$$tempscript"
+            |}
+            |trap clean_up EXIT""".stripMargin
+        val scriptPath = "\\$tempscript"
+
         s"""
-           |set -e
-           |tempscript=\\$$(mktemp "$$VIASH_TEMP/viash-run-${functionality.name}-XXXXXX")
-           |function clean_up {
-           |  rm "\\$$tempscript"
-           |}
-           |trap clean_up EXIT
-           |cat > "\\$$tempscript" << 'VIASHMAIN'
-           |$escapedCode
-           |VIASHMAIN$cdToResources$resourcesToPath
-           |${res.meta.command("\\$tempscript")}
-           |""".stripMargin
+          |set -e$scriptSetup
+          |cat > "$scriptPath" << 'VIASHMAIN'
+          |$escapedCode
+          |VIASHMAIN$cdToResources$resourcesToPath
+          |${res.meta.command(scriptPath)}
+          |""".stripMargin
+
+      // if mainResource is a script
+      case Some(res) if debugPath.isDefined =>
+        val code = res.readWithPlaceholder(functionality).get
+        val escapedCode = escapeViash(code)
+        val deb = debugPath.get
+
+        s"""
+          |set -e
+          |cat > "${debugPath.get}" << 'VIASHMAIN'
+          |$escapedCode
+          |VIASHMAIN
+          |""".stripMargin
     }
 
     // generate bash document